Indicate goldfish drivers depend on HAS_IOMEM
Goldfish drivers make use of iomem to communicate with the host.
Indicate this dependency in the Kconfig files so that drivers are not
built on platforms that do not support this. This is useful for usermode
Linux which is used to run test cases on the kernel. These drivers do
not compile in usermode without this fix because iomem is not available.

Change-Id: Ic4f9498924e46e18122ce962aa95400b69d05c2c
Signed-off-by: "Bjoern Johansson" <bjoernj@google.com>