mac80211_hwsim: add option to set unique MAC addr
Create a module parameter to set the second and third octet in the MAC
address to something unique. This allows each radio to have a globally
unique MAC address if multiple virtual wireless networks are bridged or
in the presence of other interfaces that use the same range of MAC
addresses.

Change-Id: I167ecf7a4a058a9dce2a374f620c2963f7e28588
Signed-off-by: Bjoern Johansson <bjoernj@google.com>