mac80211_hwsim: add support for packet injection
Add support for injecting packets on the hwsim monitor interface. The
injected packets are expected to have an initial radiotap header
followed by an IEEE802.11 header and payload. The radiotap header is
expected to specify which channel to transmit the frame on. Any radio
that is currently on this channel will receive the injected frame.

Additionally frames that are sent out on the monitor interface will now
be considered ACK'd in the tx_status as long as the monitor interface is
up. Otherwise userspace applications cannot work properly as their
transmissions are not considered ACK'd even though they may receive
injected responses.

Change-Id: Ieddeb41fef4481b8f15cd25921f495dcbd9fc1fe
Signed-off-by: Bjoern Johansson <bjoernj@google.com>