goldfish_address_space: Add IOCTL_PING
bug: 123884226

IOCTL_PING is a mechanism for userspace drivers to notify the host of
control/data messages. The user passes a struct with the following
parameters:

First, there are transfer related parameters:

u64 offset: The offset into any owned IOCTL_ALLOCATE_BLOCK memory. The offset
must be owned by the fd.
u64 size: The size of the transfer.

This offset is translated into an absolute physical address by the
driver, and then it is consumed in userspace.

Note that there is no guarantee that the resulting physical address is
backed by any host memory. It is up to the driver to decide how to
service the memory at that offset.

Then there is a parameter that can make it easier to access virtual
device specific context objects:

u64 metadata

This is usually either an enum saying what type of device to start up,
or a handle to or an actual host pointer.

Next are MWAIT parameters (not yet implemented).

u64 wait_offset: The offset into any owned IOCTL_ALLOCATE_BLOCK memory
where the guest should expect a reply.
u32 wait_flags: How to wait for the resulting data (poll, irq, etc)

Finally there is a single u32 controlling the direction of the transfer.
This ends up being another meta parameter that the host handles; it is
not necessary for this parameter to dictate the direction of data
transfer, it is merely a protocol convenience for any guest driver or
host device implementations.

u32 direction

Signed-off-by: Lingfeng Yang <lfy@google.com>
Change-Id: Icbccbb86b8d8bffd2f5b610478f2e8314cd1f106